+++ src/ui/main/mod.rs
@@ -451,7 +451,6 @@ impl SimpleComponent for App {
                                             Some(LauncherState::PredownloadAvailable { patch: JadeitePatchStatusVariant::Unsafe, .. }) => false,
 
                                             Some(_) => true,
-
                                             None => false
                                         },
 
@@ -466,7 +465,6 @@ impl SimpleComponent for App {
                                             Some(LauncherState::PredownloadAvailable { patch: JadeitePatchStatusVariant::Unsafe, .. }) => &["error", "pill"],
 
                                             Some(_) => &["suggested-action", "pill"],
-
                                             None => &["pill"]
                                         },
 
@@ -505,9 +503,6 @@ impl SimpleComponent for App {
                                             set_icon_name: "violence-symbolic", // window-close-symbolic
                                             set_label: &tr!("kill-game-process")
                                         },
-
-                                        #[watch]
-                                        set_visible: model.kill_game_button,
 
                                         #[watch]
                                         set_sensitive: !model.disabled_kill_game_button,
@@ -731,19 +726,28 @@ impl SimpleComponent for App {
 
                 let web_cache = config.game.path.for_edition(config.launcher.edition)
                     .join(config.launcher.edition.data_folder())
-                    .join("webCaches/Cache/Cache_Data/data_2");
+                    .join("webCaches");
 
-                if !web_cache.exists() {
-                    tracing::error!("Couldn't find wishes URL: cache file doesn't exist");
+                // Find newest cache folder
+                let mut web_cache_id = None;
 
-                    sender.input(AppMsg::Toast {
-                        title: tr!("wish-url-search-failed"),
-                        description: None
-                    });
+                if let Ok(entries) = web_cache.read_dir() {
+                    for entry in entries.flatten() {
+                        if entry.path().is_dir() &&
+                           entry.file_name().to_string_lossy().trim_matches(|c| "0123456789.".contains(c)).is_empty() &&
+                           Some(entry.file_name()) > web_cache_id
+                        {
+                            web_cache_id = Some(entry.file_name());
+                        }
+                    }
                 }
 
-                else {
-                    match std::fs::read(&web_cache) {
+                if let Some(web_cache_id) = web_cache_id {
+                    let web_cache = web_cache
+                        .join(web_cache_id)
+                        .join("Cache/Cache_Data/data_2");
+
+                    match std::fs::read(web_cache) {
                         Ok(web_cache) => {
                             let web_cache = String::from_utf8_lossy(&web_cache);
 
@@ -781,6 +785,15 @@ impl SimpleComponent for App {
                             });
                         }
                     }
+                }
+
+                else {
+                    tracing::error!("Couldn't find wishes URL: cache file doesn't exist");
+
+                    sender.input(AppMsg::Toast {
+                        title: tr!("wish-url-search-failed"),
+                        description: None
+                    });
                 }
             }));
         })));