+++ src/ui/preferences/general/mod.rs
@@ -31,7 +31,6 @@ pub struct GeneralApp {
     main_patch: Option<(Version, JadeitePatchStatusVariant)>,
 
     style: LauncherStyle,
-
     languages: Vec<String>
 }
 
@@ -200,21 +199,15 @@ impl SimpleAsyncComponent for GeneralApp {
                         &tr!("china")
                     ])),
 
-                    set_selected: match CONFIG.launcher.edition {
-                        GameEdition::Global => 0,
-                        GameEdition::China => 1
-                    },
+                    set_selected: GameEdition::list().iter()
+                        .position(|edition| edition == &CONFIG.launcher.edition)
+                        .unwrap() as u32,
 
                     connect_selected_notify[sender] => move |row| {
                         if is_ready() {
                             #[allow(unused_must_use)]
                             if let Ok(mut config) = Config::get() {
-                                config.launcher.edition = match row.selected() {
-                                    0 => GameEdition::Global,
-                                    1 => GameEdition::China,
-
-                                    _ => unreachable!()
-                                };
+                                config.launcher.edition = GameEdition::list()[row.selected() as usize];
 
                                 Config::update(config);
 
@@ -471,7 +464,6 @@ impl SimpleAsyncComponent for GeneralApp {
             main_patch: None,
 
             style: CONFIG.launcher.style,
-
             languages: SUPPORTED_LANGUAGES.iter().map(|lang| tr!(format_lang(lang).as_str())).collect()
         };